[Hexagon] Refactor Hexagon.cmake (#10227)
This file is included every time TVM is build, regardless of whether
any support for Hexagon is enabled or not. This refactoring is meant
to remove underlying assumptions about what features are enabled and
what the compilation targets are. Now, when there is nothing needed
from Hexagon, the script exits early (although it doesn't need to),
and the rest of it is (and should remain) safe to execute regardless
of build configuration.

Disable "runtime.module.loadfile_hexagon" from the offload runtime,
since it conflicts with device_api.hexagon.v2.
It was only used with offload on Android, which is being deprecated.